chore: move getTextEditor
